% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_data}
\alias{aggregate_data}
\alias{dw_aggregate_data}
\title{Aggregate Data with Optional Weighting and Global Aggregation}
\usage{
aggregate_data(
  data,
  value,
  weight,
  by,
  global = TRUE,
  method = c("mean", "weighted_mean"),
  pop.coverage = FALSE,
  country.coverage = FALSE
)

dw_aggregate_data(
  data,
  value,
  weight,
  by,
  global = TRUE,
  method = c("mean", "weighted_mean"),
  pop.coverage = FALSE,
  country.coverage = FALSE
)
}
\arguments{
\item{data}{A data frame containing the data to be aggregated.}

\item{value}{The name of the column in `data` that contains the values to aggregate.}

\item{weight}{The name of the column in `data` that contains the weights for weighted aggregation.}

\item{by}{A character vector of column names in `data` to group by for aggregation.}

\item{global}{Logical; if TRUE, includes a global aggregation ignoring the `by` groups.}

\item{method}{The method of aggregation; either "mean" for unweighted mean or "weighted_mean" for weighted mean.}

\item{pop.coverage}{Logical; if TRUE, shows share of child population in the reference group covered with data}

\item{country.coverage}{Logical; if TRUE, adds a column counting the number of non-NA observations contributing to each group.}
}
\value{
A data frame with aggregated data.
}
\description{
This function aggregates data by specified grouping variables, with options for
weighted aggregation and global aggregation. It calculates coverage of non-NA
values and allows for mean or weighted mean aggregation methods.
}
\examples{
\dontrun{
library(dplyr)
library(tidyr)
data <- data.frame(
  region = rep(c("North", "South"), each = 5),
  value  = c(1:10),
  weight = c(5:14)
)
aggregate_data(data, value = "value", weight = "weight", by = "region",
               global = TRUE, method = "mean")
}
}
\seealso{
[aggregate_data_v2()] for the enhanced cross-sector variant;
  [generate_agg_footnote()] for the standard footnote string.

Other aggregate: 
\code{\link{aggregate_data_v2}()},
\code{\link{apply_time_window}()},
\code{\link{generate_agg_footnote}()}
}
\concept{aggregate}
